standalone demo of Data::Dumper problem
[needless to say, this branch will be deleted eventually...]

Please see comments in the data-dumper-returns-undef.pl program to
understand what is happening.  Typically anyone with large
configurations see this happening.

I have worked around it by coding around the need for a custom sort sub,
and as you can see there is at least one other way to make it work (look
for WTF in the code!) but those are just work arounds...